requirements.txt
# Purpose: Your Python "Gemfile".


fastapi
uvicorn
anthropic
python-dotenv
pydantic
requests